Moved use case definitions to separate files.
