+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/complain.js
@@ -1,63 +1,12 @@
-/**
- * Created by fsuda on 18.09.2018.
- */
-import { details, emptyDraft } from "../detail-definitions.js";
-import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
+import { complain } from "./uc-complain";
+import { handleComplaint } from "./uc-handle-complaint";
 
 export const complainGroup = {
   identifier: "complaingroup",
   label: "Complaints",
   icon: undefined,
   useCases: {
-    complain: {
-      identifier: "complain",
-      label: "Complain about something",
-      icon: "#ico36_uc_wtf",
-      timeToLiveMillisDefault: 1000 * 60 * 60 * 24 * 30,
-      doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-      draft: {
-        ...emptyDraft,
-        content: {
-          type: "won:Complaint",
-          title: "WTF?",
-          tags: ["wtf"],
-        },
-        seeks: {
-          type: "won:HandleComplaint",
-        },
-      },
-      details: {
-        title: { ...details.title },
-        description: { ...details.description },
-        location: { ...details.location },
-        tags: { ...details.tags },
-      },
-      seeksDetails: undefined,
-    },
-    handleComplaints: {
-      identifier: "handleComplaints",
-      label: "Handle complaints",
-      icon: "#ico36_uc_wtf_interest",
-      timeToLiveMillisDefault: 1000 * 60 * 60 * 24 * 30,
-      doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-      draft: {
-        ...emptyDraft,
-        content: {
-          type: "won:HandleComplaint",
-          title: "I'll discuss complaints",
-          searchString: "wtf",
-        },
-        seeks: {
-          type: "won:Complaint",
-        },
-      },
-      details: {
-        title: { ...details.title },
-        description: { ...details.description },
-        location: { ...details.location },
-        tags: { ...details.tags },
-      },
-      seeksDetails: undefined,
-    },
+    complain: complain,
+    handleComplaint: handleComplaint,
   },
 };
