+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-afterparty.js
@@ -1,0 +1,29 @@
+/**
+ * Created by fsuda on 18.09.2018.
+ */
+import { details, emptyDraft } from "../detail-definitions.js";
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
+import { interestsDetail } from "../details/person.js";
+
+export const afterparty = {
+  identifier: "afterparty",
+  label: "Go out",
+  icon: "#ico36_uc_drinks",
+  doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  draft: {
+    ...emptyDraft,
+    content: {
+      title: "I'm up for partying! Any plans?",
+      tags: ["afterparty"],
+      searchString: "afterparty",
+    },
+  },
+  details: {
+    title: { ...details.title },
+    fromDatetime: { ...details.fromDatetime },
+    throughDatetime: { ...details.throughDatetime },
+    description: { ...details.description },
+    location: { ...details.location },
+    interests: { ...interestsDetail },
+  },
+};
