+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-custom.js
@@ -1,0 +1,14 @@
+import { details, emptyDraft } from "../detail-definitions.js";
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
+
+export const custom = {
+  custom: {
+    identifier: "allDetails",
+    label: "New custom post",
+    icon: "#ico36_uc_custom",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: { ...emptyDraft },
+    details: details,
+    seeksDetails: details,
+  },
+};
