+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-get-lunch.js
@@ -1,0 +1,30 @@
+/**
+ * Created by fsuda on 18.09.2018.
+ */
+import { details, emptyDraft } from "../detail-definitions.js";
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
+import { interestsDetail } from "../details/person.js";
+
+export const getLunch = {
+  identifier: "lunch",
+  label: "Get lunch",
+  icon: "#ico36_uc_meal-half",
+  doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  draft: {
+    ...emptyDraft,
+    content: {
+      title: "I'm up for lunch! Any plans?",
+      tags: ["lunch"],
+      searchString: "lunch",
+    },
+  },
+  details: {
+    title: { ...details.title },
+    description: { ...details.description },
+    fromDatetime: { ...details.fromDatetime },
+    throughDatetime: { ...details.throughDatetime },
+    location: { ...details.location },
+    interests: { ...interestsDetail },
+  },
+  seeksDetails: undefined,
+};
