+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-postdoc-offer.js
@@ -1,0 +1,30 @@
+/**
+ * Created by fsuda on 18.09.2018.
+ */
+import { details, emptyDraft } from "../detail-definitions.js";
+import { interestsDetail, skillsDetail } from "../details/person.js";
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
+
+export const postdocOffer = {
+  identifier: "postDocIs",
+  label: "Offer a PostDoc position",
+  icon: "#ico36_uc_postdoc",
+  doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  draft: {
+    ...emptyDraft,
+    content: {
+      title: "I'm offering a PostDoc position!",
+      tags: ["offer-postdoc"],
+      searchString: "search-postdoc",
+    },
+  },
+  details: {
+    title: { ...details.title },
+    description: { ...details.description },
+    location: { ...details.location },
+  },
+  seeksDetails: {
+    skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
+    interests: { ...interestsDetail, placeholder: "" },
+  },
+};
